[pvr] Refactor: Move CPVRGUIActionsPlayback::StartPlayback to class CPVRPlaybackState.
